rename geography to geography-author
