+++ bin/mdsctl
@@ -137,7 +137,7 @@ declare -A apps=(
   ["mds-audit"]="https://localhost:[port]/audit,mds,app"
   ["mds-compliance"]="https://localhost:[port]/compliance,mds,app"
   ["mds-daily"]="https://localhost:[port]/daily,mds,app"
-  ["mds-geography"]="https://localhost:[port]/geography,mds,app"
+  ["mds-geography-author"]="https://localhost:[port]/geography-author,mds,app"
   ["mds-native"]="https://localhost:[port]/native,mds,app"
   ["mds-policy"]="https://localhost:[port]/policy,mds,app"
   ["mds-policy-author"]="https://localhost:[port]/policy-author,mds,app"
@@ -264,7 +264,7 @@ where app in:
   mds-audit
   mds-compliance
   mds-daily
-  mds-geography
+  mds-geography-author
   mds-native
   mds-policy
   mds-policy-author
