+++ container-images/mds-geography-author/Dockerfile
