+++ container-images/mds-geography-author/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@container-images/mds-geography",
+  "name": "@container-images/mds-geography-author",
   "version": "0.0.1",
   "description": "Docker Image for MDS Geography API",
   "main": "dist/index.js",
@@ -22,7 +22,7 @@
   "dependencies": {
     "@container-images/env-inject": "0.1.25",
     "@mds-core/mds-api-server": "0.1.26",
-    "@mds-core/mds-geography": "0.0.1"
+    "@mds-core/mds-geography-author": "0.0.1"
   },
   "license": "Apache-2.0"
 }
