+++ container-images/mds-geography-author/tsconfig.build.json
@@ -8,7 +8,7 @@
       "path": "../../container-images/env-inject/tsconfig.build.json"
     },
     {
-      "path": "../../packages/mds-geography/tsconfig.build.json"
+      "path": "../../packages/mds-geography-author/tsconfig.build.json"
     },
     {
       "path": "../../packages/mds-api-server/tsconfig.build.json"
