+++ container-images/mds-geography-author/webpack.config.js
