+++ packages/mds-geography-author/README.md
@@ -1,2 +1,2 @@
-# mds-geography
+# mds-geography-author
 Read and write for MDS Geography objects
