+++ packages/mds-geography-author/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@mds-core/mds-geography",
-  "description": "MDS Geography API",
+  "name": "@mds-core/mds-geography-author",
+  "description": "MDS Geography Author API",
   "version": "0.0.1",
   "private": true,
   "licence": "Apache-2.0",
@@ -25,11 +25,11 @@
   "types": "dist/index.d.ts",
   "scripts": {
     "build": "tsc --build tsconfig.build.json",
-    "start": "PATH_PREFIX=/geography yarn watch server",
+    "start": "PATH_PREFIX=/geography-author yarn watch server",
     "test": "yarn test:prettier && yarn test:eslint && yarn test:unit",
     "test:eslint": "eslint --fix --ignore-path ../../.gitignore '**/*.ts'",
     "test:prettier": "prettier --write --check --ignore-path ../../.gitignore '**/*.ts'",
-    "test:unit": "PATH_PREFIX=/geography DOTENV_CONFIG_PATH=../../.env nyc --lines 85 ts-mocha --project ../../tsconfig.json",
+    "test:unit": "PATH_PREFIX=/geography-author DOTENV_CONFIG_PATH=../../.env nyc --lines 85 ts-mocha --project ../../tsconfig.json",
     "watch": "nodemon --watch '../../packages' --ext 'ts' --ignore '*.d.ts' --exec yarn watch:exec --",
     "watch:exec": "yarn build && ts-node"
   }
